requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.6.2; python_full_version >= '3.6.2'
asgiref==3.6.0; python_version >= '3.7'
async-timeout==4.0.2; python_version >= '3.6'
backoff==2.2.1; python_version >= '3.7'
certifi==2022.12.7; python_version >= '3.6'
charset-normalizer==3.0.1
click==8.1.3; python_version >= '3.7'
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dnspython==2.3.0; python_version >= '3.7' and python_version < '4'
fastapi==0.92.0
flask==2.2.3
googleapis-common-protos==1.56.2; python_version >= '3.6'
grpcio==1.51.3; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==0.16.3; python_version >= '3.7'
httpx==0.23.3
idna==3.4; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.2; python_version >= '3.7'
opentelemetry-api==1.16.0; python_version >= '3.7'
opentelemetry-exporter-jaeger-proto-grpc==1.16.0; python_version >= '3.7'
opentelemetry-exporter-jaeger-thrift==1.16.0; python_version >= '3.7'
opentelemetry-exporter-jaeger==1.16.0
opentelemetry-exporter-otlp-proto-grpc==1.16.0; python_version >= '3.7'
opentelemetry-exporter-otlp-proto-http==1.16.0; python_version >= '3.7'
opentelemetry-exporter-otlp==1.16.0
opentelemetry-instrumentation-asgi==0.37b0; python_version >= '3.7'
opentelemetry-instrumentation-fastapi==0.37b0
opentelemetry-instrumentation-flask==0.37b0
opentelemetry-instrumentation-httpx==0.37b0
opentelemetry-instrumentation-pymongo==0.37b0
opentelemetry-instrumentation-redis==0.37b0
opentelemetry-instrumentation-requests==0.37b0
opentelemetry-instrumentation-wsgi==0.37b0; python_version >= '3.7'
opentelemetry-instrumentation==0.37b0; python_version >= '3.7'
opentelemetry-proto==1.16.0; python_version >= '3.7'
opentelemetry-sdk==1.16.0; python_version >= '3.7'
opentelemetry-semantic-conventions==0.37b0; python_version >= '3.7'
opentelemetry-util-http==0.37b0; python_version >= '3.7'
protobuf==3.20.3; python_version >= '3.7'
pydantic==1.10.5; python_version >= '3.7'
pymongo==4.3.3
redis==4.5.1
requests==2.28.2
rfc3986[idna2008]==1.5.0
setuptools==67.4.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.0; python_version >= '3.7'
starlette==0.25.0; python_version >= '3.7'
thrift==0.16.0
typing-extensions==4.5.0; python_version >= '3.7'
urllib3==1.26.14; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn==0.20.0
werkzeug==2.2.3; python_version >= '3.7'
wrapt==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
